Move states test to lib.rs
