+++ crates/bevy_ecs/src/lib.rs
@@ -21,6 +21,7 @@ pub mod world;
 
 use std::any::TypeId;
 
+use bevy_ecs_macros::States;
 pub use bevy_ptr as ptr;
 
 /// Most commonly used re-exported types.
@@ -1707,4 +1708,38 @@ mod tests {
             "new entity was spawned and received C component"
         );
     }
+
+    #[test]
+    fn fieldful_and_fieldless_states() {
+        #[derive(Hash, Eq, PartialEq, Clone, Debug, States)]
+        pub enum Foo {
+            Fieldless,
+            Fieldful(Bar),
+        }
+        #[derive(Hash, Eq, PartialEq, Clone, Debug, States)]
+        pub enum Bar {
+            Alice,
+            Bob,
+        }
+        impl Default for Bar {
+            fn default() -> Self {
+                Self::Alice
+            }
+        }
+        impl Default for Foo {
+            fn default() -> Self {
+                Self::Fieldless
+            }
+        }
+
+        assert_eq!(
+            Foo::variants().collect::<Vec<Foo>>(),
+            vec![
+                Foo::Fieldless,
+                Foo::Fieldful(Bar::Alice),
+                Foo::Fieldful(Bar::Bob)
+            ]
+        )
+    }
+
 }
